+++ src/main/java/org/opensearch/knn/index/KNNMethodContext.java
@@ -65,8 +65,6 @@ public class KNNMethodContext implements ToXContentFragment, Writeable {
     @NonNull
     private final MethodComponentContext methodComponentContext;
 
-    private static final String DELIMITER = ";";
-
     /**
      * Constructor from stream.
      *